requirements.txt
Django == 4.1.2
djangorestframework == 3.14.0
